Review comments: moved the tabs creation to utility function
