+++ src/client/flogo/flow/triggers/configurator/core/constants.ts
@@ -1,0 +1,4 @@
+export const TRIGGER_TABS = {
+  MAP_FLOW_INPUT: 'mapFlowInput',
+  MAP_FLOW_OUTPUT: 'mapFlowOutput'
+};
